+++ lib/spectre.rb
@@ -13,10 +13,6 @@ module Spectre
       merger = proc { |_key, v1, v2| Hash === v1 && Hash === v2 ? v1.merge!(v2, &merger) : v2 }
       self.merge!(second, &merger)
     end
-
-    def deep_clone
-      Marshal.load(Marshal.dump(self))
-    end
   end
 
   class ::Object
@@ -24,6 +20,10 @@ module Spectre
       self.instance_variables.each_with_object({}) do |var, hash|
         hash[var.to_s.delete("@")] = self.instance_variable_get(var)
       end
+    end
+
+    def deep_clone
+      Marshal.load(Marshal.dump(self))
     end
   end
 
@@ -84,16 +84,24 @@ module Spectre
       @name = desc.downcase.gsub(/[^a-z0-9]+/, '_')
     end
 
-    def add_spec desc, tags, data, block, context, file
+    def add_spec desc, tags, data, block, context, file, line
       name = @name + '-' + (@specs.length+1).to_s
-      @specs << Spec.new(name, self, desc, tags, data, block, context, file)
+      @specs << Spec.new(name, self, desc, tags, data, block, context, file, line)
+    end
+
+    def to_h
+      {
+        name: @name,
+        desc: @desc,
+        specs: @specs.map { |x| x.to_h }
+      }
     end
   end
 
   class Spec
-    attr_reader :name, :subject, :context, :desc, :tags, :data, :block, :file
+    attr_reader :id, :name, :subject, :context, :desc, :tags, :data, :block, :file, :line
 
-    def initialize name, subject, desc, tags, data, block, context, file
+    def initialize name, subject, desc, tags, data, block, context, file, line
       @name = name
       @context = context
       @data = data
@@ -102,10 +110,24 @@ module Spectre
       @tags = tags
       @block = block
       @file = file
+      @line = line
     end
 
     def full_desc
       @subject.desc + ' ' + desc
+    end
+
+    def to_h
+      {
+        name: @name,
+        context: @context.__desc,
+        data: @data.map { |x| x.to_h },
+        subject: @subject.desc,
+        desc: @desc,
+        tags: @tags,
+        file: @file,
+        line: @line,
+      }
     end
   end
 
@@ -149,22 +171,41 @@ module Spectre
 
       return :success
     end
+
+    def to_h
+      date_format = '%FT%T.%L%:z'
+
+      {
+        spec: @spec.name,
+        data: @data,
+        started: @started.strftime(date_format),
+        finished: @finished.strftime(date_format),
+        error: @error,
+        failure: @failure,
+        skipped: @skipped,
+        log: @log.map { |timestamp, message, level, name| [timestamp.strftime(date_format), message, level, name] },
+        expectations: @expectations,
+        properties: @properties,
+      }
+    end
   end
 
   class Runner
-    @@current = nil
-
     def self.current
-      @@current
+      Thread.current.thread_variable_get('current_run')
+    end
+
+    def self.current= run
+      Thread.current.thread_variable_set('current_run', run)
     end
 
     def run specs
       runs = []
 
       specs.group_by { |x| x.subject }.each do |subject, subject_specs|
-        Logging.log_subject subject do
+        Spectre::Logging.log_subject subject do
           subject_specs.group_by { |x| x.context }.each do |context, context_specs|
-            Logging.log_context(context) do
+            Spectre::Logging.log_context(context) do
               runs.concat run_context(context, context_specs)
             end
           end
@@ -193,12 +234,12 @@ module Spectre
             spec.data
               .map { |x| x.is_a?(Hash) ? OpenStruct.new(x) : x }
               .each do |data|
-                Logging.log_spec(spec, data) do
+                Spectre::Logging.log_spec(spec, data) do
                   runs << run_spec(spec, data)
                 end
               end
           else
-            Logging.log_spec(spec) do
+            Spectre::Logging.log_spec(spec) do
               runs << run_spec(spec)
             end
           end
@@ -215,11 +256,11 @@ module Spectre
     def run_setup spec
       run_info = RunInfo.new(spec)
 
-      @@current = run_info
+      Runner.current = run_info
 
       run_info.started = Time.now
 
-      Logging.log_context(spec.context) do
+      Spectre::Logging.log_context(spec.context) do
         begin
           spec.block.call()
 
@@ -228,13 +269,13 @@ module Spectre
           run_info.failure = e
         rescue Exception => e
           run_info.error = e
-          Logging.log_error(spec, e)
+          Spectre::Logging.log_error(spec, e)
         end
       end
 
       run_info.finished = Time.now
 
-      @@current = nil
+      Runner.current = nil
 
       run_info
     end
@@ -242,7 +283,7 @@ module Spectre
     def run_spec spec, data=nil
       run_info = RunInfo.new(spec, data)
 
-      @@current = run_info
+      Runner.current = run_info
 
       run_info.started = Time.now
 
@@ -250,7 +291,7 @@ module Spectre
         if spec.context.__before_blocks.count > 0
           before_ctx = SpecContext.new(spec.subject, 'before', spec.context)
 
-          Logging.log_context before_ctx do
+          Spectre::Logging.log_context before_ctx do
             spec.context.__before_blocks.each do |block|
               block.call(data)
             end
@@ -262,18 +303,18 @@ module Spectre
         run_info.failure = e
       rescue SpectreSkip => e
         run_info.skipped = true
-        Logging.log_skipped(spec, e.message)
+        Spectre::Logging.log_skipped(spec, e.message)
       rescue Interrupt
         run_info.skipped = true
-        Logging.log_skipped(spec, 'canceled by user')
+        Spectre::Logging.log_skipped(spec, 'canceled by user')
       rescue Exception => e
         run_info.error = e
-        Logging.log_error(spec, e)
+        Spectre::Logging.log_error(spec, e)
       ensure
         if spec.context.__after_blocks.count > 0
           after_ctx = SpecContext.new(spec.subject, 'after', spec.context)
 
-          Logging.log_context after_ctx do
+          Spectre::Logging.log_context after_ctx do
             begin
               spec.context.__after_blocks.each do |block|
                 block.call
@@ -284,7 +325,7 @@ module Spectre
               run_info.failure = e
             rescue Exception => e
               run_info.error = e
-              Logging.log_error(spec, e)
+              Spectre::Logging.log_error(spec, e)
             end
           end
         end
@@ -292,7 +333,7 @@ module Spectre
 
       run_info.finished = Time.now
 
-      @@current = nil
+      Runner.current = nil
 
       run_info
     end
@@ -319,9 +360,9 @@ module Spectre
     end
 
     def it desc, tags: [], with: [], &block
-      spec_file = get_file()
+      spec_file, line = get_call_location()
 
-      @__subject.add_spec(desc, tags, with, block, self, spec_file)
+      @__subject.add_spec(desc, tags, with, block, self, spec_file, line)
     end
 
     def before &block
@@ -333,19 +374,19 @@ module Spectre
     end
 
     def setup &block
-      name = "#{@__subject.name}-setup"
-      spec_file = get_file()
+      name = "#{@__subject.name}-setup-#{@__setup_blocks.count+1}"
+      spec_file, line = get_call_location()
 
       setup_ctx = SpecContext.new(@__subject, 'setup', self)
-      @__setup_blocks << Spec.new(name, @__subject, 'setup', [], nil, block, setup_ctx, spec_file)
+      @__setup_blocks << Spec.new(name, @__subject, 'setup', [], nil, block, setup_ctx, spec_file, line)
     end
 
     def teardown &block
-      name = "#{@__subject.name}-teardown"
-      spec_file = get_file()
+      name = "#{@__subject.name}-teardown-#{@__teardown_blocks.count+1}"
+      spec_file, line = get_call_location()
 
       teardown_ctx = SpecContext.new(@__subject, 'teardown', self)
-      @__teardown_blocks << Spec.new(name, @__subject, 'teardown', [], nil, block, teardown_ctx, spec_file)
+      @__teardown_blocks << Spec.new(name, @__subject, 'teardown', [], nil, block, teardown_ctx, spec_file, line)
     end
 
     def context desc=nil, &block
@@ -355,21 +396,11 @@ module Spectre
 
     private
 
-    def get_file
-      # Get the file, where the spec is defined.
-      # Nasty, but it works
-      # Maybe there is another way, but this works for now
-
-      begin
-        raise
-      rescue => e
-        return e.backtrace
-          .select { |file| !file.include? 'lib/spectre' }
-          .first
-          .match(/(.*\.rb):\d+/)
-          .captures
-          .first
-      end
+    def get_call_location
+      path_and_line = caller[1].split(':')
+      line = path_and_line[-2].to_i
+      file = path_and_line[0..-3].join(':')
+      [file, line]
     end
   end
 
@@ -409,8 +440,9 @@ module Spectre
     @@subjects = []
     @@modules = []
 
-    attr_reader :file_log, :logger
-
+    def subjects
+      @@subjects
+    end
 
     def specs spec_filter=[], tags=[]
       @@subjects
@@ -443,6 +475,11 @@ module Spectre
       end
     end
 
+    def purge
+      @@subjects = []
+      @@modules = []
+    end
+
 
     ###########################################
     # Global Functions
@@ -470,7 +507,7 @@ module Spectre
     end
   end
 
-  delegate :describe, :property, :skip, to: Spectre
+  delegate(:describe, :property, :skip, to: Spectre)
 end